backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
faiss-cpu
transformers
Pillow
pandas
scikit-learn
numpy
pinecone
langchain
huggingface_hub
joblib
requests
